SCRD-2020 Use global application group
Forcing the global application group is recommended for
daemonified wsgi applications and ensures the application
will always be executed within the context of the first
interpreter created by Python when it is initialised.

Forcing a WSGI application to run within the first
interpreter can be necessary when a third party C extension
module for Python has used the simplified threading API
for manipulation of the Python GIL and thus will not
run correctly within any additional sub interpreters
created by Python.

Change-Id: I487f658918970ea99de0b24fe15473989b00aa71