SCRD-2756 don't install systemd service files in SUSEified deployment
In a SUSEified deployment, systemd service files are installed by the
rpms, so stop Ardana from installing extra ones into /etc which
override them.

Later we may need to go through the differences in more detail to
check we aren't missing anything in the rpms which the *.service.j2
templates offer, but in the case of the ELK stack, the rpms are
already tested in a Crowbar context and known to be good, so it should
be fairly safe to use the rpm versions.

On the other side, here are a couple of specific examples of why the
current *.service.j2 files won't work well:

1. If we're installing kibana from an rpm, we need to trust the rpm to
   handle file ownership correctly, otherwise the whole rpm
   verification process becomes useless.  So we need to avoid

     ExecStartPre=/bin/chown -R kibana.root {{ kibana_path }}

2. elasticsearch.service.j2 has

     EnvironmentFile=/etc/default/elasticsearch

   but in the SUSE world this needs to be /etc/sysconfig.

Change-Id: I166274bdb3620581b3f33d54dc71594b36898706