Resolves beaver permission issues (bsc#1079673)
Chmods o+rx to allow the beaver user access to its configuration files.
The alternative is to change the directory's group from root to
adm but the former option is chosen to match legacy installs.

Adds the beaver user to the adm group to provide read access to
openstack service log files.

Change-Id: I853ab17726e95336bb9165021608bd561065a24f