Updates prune_es_indices.sh to v4.2 (bsc#1092986)
Updates prune_es_indices.sh to elasticsearch's curator version 4.2.

Replaces curator with curator_cli. The current version of the curator
module no longer accepts raw parameters. It requires a connection file
and an action file and it no longer provides the show indices action.
There is a new helper module named curator_cli which behaves similar
to the old curator module and it provides the show indices action.

Removes disabling of bloom filters because according to the
documentation "Beginning with version 1.4, Elasticsearch no longer
enables bloom filters for search" and we ship version 2.4.

Change-Id: Icaeef5f5c5506cd697d67ee73c82eb9279bc64b9