SCRD-5366 Avoid rewriting ardana service trigger file
As part of the log cleanup, this job was compressing the existing
trigger file and writing a new one, which triggered the ardana service
to abruptly restart, even if it was in the middle of running a playbook.
This change removes the trigger file from the set of files that are
automatically compressed and re-created nightly.

Change-Id: Ieb6be4aa707fd72adf885158f412e8288ba3defe