SCRD-5855 Remove freezer playbook references in kronos playbooks
The freezer service is being removed in Cloud 9, so we need to remove
the references to the freezer playbooks that exist in other service
playbooks, such as those in the kronos management playbooks in the
logging-ansible repo.

However as these are the only actions in these kronos playbooks, we
can just remove these playbooks, and the references to them in the
logging-upgrade.yml playbook.

Change-Id: Ia12a00f46d3e09aeb2b478c11effc9019f483da6